WIP: extracting location service
