+++ tempted/Constants.swift
@@ -49,4 +49,10 @@ struct TPTError {
         NSLocalizedFailureReasonErrorKey: NSLocalizedString("Permissions not granted.", comment: "internal error reason for no permissions"),
         NSLocalizedRecoverySuggestionErrorKey: NSLocalizedString("Please grant photo permissions in your settings", comment: "User-facing recovery suggestions")
     ])
+    
+    static let MapNoPermissions = NSError(domain: "tempted", code: 67, userInfo: [
+        NSLocalizedDescriptionKey: NSLocalizedString("Error getting latlng", comment: "internal error description for capturing location"),
+        NSLocalizedFailureReasonErrorKey: NSLocalizedString("Permissions not granted.", comment: "internal error reason for no permissions"),
+        NSLocalizedRecoverySuggestionErrorKey: NSLocalizedString("Please grant location permissions in your settings", comment: "User-facing recovery suggestions")
+    ])
 }
